+++ CHANGES.md
@@ -5,3 +5,6 @@
 ## v0.0.0 - 2024-xx-xx
 - Implement `linksmith inventory` and `linksmith output-formats`
   subcommands, based on `sphobjinv` and others. Thanks, @bskinn.
+- Implement `linksmith anansi suggest`, also available as `anansi`,
+  to easily suggest terms of a few curated community projects.
+  Thanks, @bskinn.